# fixtures/golden/codegen/go/chi/mysql/url_shortener/openapi.yaml
openapi: 3.1.0
info:
  title: UrlShortener
  version: 0.1.0
  description: API for UrlShortener. Generated from formal specification.
servers:
- url: http://localhost:8080
  description: Local development
paths:
...
  /admin/state:
    get:
      operationId: admin_state
      summary: Export spec-level state
      description: Returns the service state projected onto the spec's state fields.
      tags:
      - admin
      responses:
        '200':
          description: State projection
          content:
            application/json:
              schema:
                type:
                - object
                additionalProperties: true
        '401':
          description: Missing or invalid admin credential
        '404':
          description: Admin surface disabled (no ADMIN_TOKEN configured on the service)
      security:
      - AdminBearer: []
  /admin/reset:
    post:
      operationId: admin_reset
      summary: Re-initialize service state
      description: Deletes all entity rows and restores scalar state fields to their invariant-derived
        seeds.
      tags:
      - admin
      responses:
        '204':
          description: State re-initialized
        '401':
          description: Missing or invalid admin credential
        '404':
          description: Admin surface disabled (no ADMIN_TOKEN configured on the service)
      security:
      - AdminBearer: []
...
components:
...
  securitySchemes:
    AdminBearer:
      type: http
      scheme: bearer
      description: Admin surface credential (ADMIN_TOKEN). With no token configured on the service, /admin
        routes answer 404.
...