# modules/codegen/src/main/scala/specrest/codegen/go/AdminRouterGo.scala
package specrest.codegen.go

import specrest.codegen.AdminModel
import specrest.convention.ScalarState
import specrest.ir.Naming
import specrest.ir.generated.SpecRestGenerated.entFields
import specrest.ir.generated.SpecRestGenerated.entName
import specrest.ir.generated.SpecRestGenerated.entity_decl
import specrest.ir.generated.SpecRestGenerated.fldName
import specrest.ir.generated.SpecRestGenerated.irStateFields
import specrest.ir.generated.SpecRestGenerated.stfName
import specrest.ir.generated.SpecRestGenerated.svcEntities
import specrest.ir.generated.SpecRestGenerated.svcTransitions
import specrest.ir.generated.SpecRestGenerated.trnEntity
import specrest.profile.ProfiledService

// Spec-derived /admin router for go-chi, compiled into every build; main.go
// passes the bearer guard (auth.RequireAdmin) so the router stays free of
// module-path knowledge. Uses generic database/sql (bun.DB embeds *sql.DB) so
// no generated models are imported. Mirrors the reset / state / seed contract
// of AdminRouter (Python) and AdminRouterTs.
object AdminRouterGo:

  def emit(profiled: ProfiledService): String =
    val ir       = profiled.ir
    val entities = svcEntities(ir)
    val isPg     = profiled.profile.database == "postgres"

    def tbl(e: entity_decl): String = Naming.toTableName(entName(e))

    val scalarFields = ScalarState.fields(ir)
    val scalarReset =
      if scalarFields.isEmpty then ""
      else
        val sets = ScalarState
          .fieldsWithSeeds(ir)
          .map((sf, seed) => s"${ScalarState.columnName(stfName(sf))} = $seed")
          .mkString(", ")
        s"""\n\t\tif _, err := db.ExecContext(ctx, ${GoLit.str(
            s"UPDATE ${ScalarState.TableName} SET $sets WHERE id = 1"
          )}); err != nil {\n\t\t\thttp.Error(w, err.Error(), 500)\n\t\t\treturn\n\t\t}"""
    val resetStmts =
      if entities.isEmpty && scalarFields.isEmpty then "\t\t// no entities"
      else
        entities.reverse
          .map(e =>
            s"""\t\tif _, err := db.ExecContext(ctx, ${GoLit.str(
                s"DELETE FROM ${tbl(e)}"
              )}); err != nil {\n\t\t\thttp.Error(w, err.Error(), 500)\n\t\t\treturn\n\t\t}"""
          )
          .mkString("\n") + scalarReset

    val stateFields = irStateFields(ir)
    val stateAssigns = stateFields
      .map: f =>
        AdminModel.projectionFor(f, ir) match
          case Some(p) =>
            p.valueShape match
              case AdminModel.ProjectionValue.ScalarStateColumn(col) =>
                s"""\t\t{
                   |\t\t\trows, err := queryAll(ctx, db, ${GoLit.str(
                    s"SELECT $col FROM ${ScalarState.TableName} WHERE id = 1"
                  )})
                   |\t\t\tif err != nil {
                   |\t\t\t\thttp.Error(w, err.Error(), 500)
                   |\t\t\t\treturn
                   |\t\t\t}
                   |\t\t\tif len(rows) > 0 {
                   |\t\t\t\tstate[${GoLit.str(stfName(f))}] = rows[0][${GoLit.str(col)}]
                   |\t\t\t} else {
                   |\t\t\t\tstate[${GoLit.str(stfName(f))}] = nil
                   |\t\t\t}
                   |\t\t}""".stripMargin
              case shape =>
                val keyCol = Naming.toColumnName(p.keyFieldName)
                val valExpr = shape match
                  case AdminModel.ProjectionValue.PrimitiveField(name) =>
                    s"row[${GoLit.str(Naming.toColumnName(name))}]"
                  case _ => "row"
                val srcTbl = entities
                  .find(e => entName(e) == p.entityName)
                  .map(tbl)
                  .getOrElse(Naming.toTableName(p.entityName))
                s"""\t\t{
                   |\t\t\trows, err := queryAll(ctx, db, ${GoLit.str(s"SELECT * FROM $srcTbl")})
                   |\t\t\tif err != nil {
                   |\t\t\t\thttp.Error(w, err.Error(), 500)
                   |\t\t\t\treturn
                   |\t\t\t}
                   |\t\t\tm := map[string]any{}
                   |\t\t\tfor _, row := range rows {
                   |\t\t\t\tm[fmt.Sprintf("%v", row[${GoLit.str(keyCol)}])] = $valExpr
                   |\t\t\t}
                   |\t\t\tstate[${GoLit.str(stfName(f))}] = m
                   |\t\t}""".stripMargin
          case None =>
            s"\t\tstate[${GoLit.str(stfName(f))}] = nil"
      .mkString("\n")

    val seedEntities = svcTransitions(ir).map(trnEntity).toSet
    val seedTargets  = entities.filter(e => seedEntities.contains(entName(e)))
    val seedHandlers = seedTargets
      .map(e => seedHandler(e, isPg))
      .mkString("\n\n")
    val seedSection = if seedTargets.isEmpty then "" else "\n\n" + seedHandlers

    s"""|// Auto-generated by spec-to-rest. Do not edit by hand.
        |//
        |// Bearer-guarded admin surface (state export / data import / re-initialize).
        |// The conformance suite (tests/) is its primary client; the contract
        |// (reset / state / seed) is identical across the fastapi/express targets.
        |
        |package admin
        |
        |import (
        |\t"context"
        |\t"encoding/json"
        |\t"fmt"
        |\t"net/http"
        |\t"strings"
        |\t"time"
        |
        |\t"github.com/go-chi/chi/v5"
        |\t"github.com/uptrace/bun"
        |)
        |
        |func join(xs []string) string { return strings.Join(xs, ", ") }
        |
        |func placeholders(n int, pg bool) string {
        |\tps := make([]string, n)
        |\tfor i := 0; i < n; i++ {
        |\t\tif pg {
        |\t\t\tps[i] = fmt.Sprintf("$$%d", i+1)
        |\t\t} else {
        |\t\t\tps[i] = "?"
        |\t\t}
        |\t}
        |\treturn strings.Join(ps, ", ")
        |}
        |
        |func normalize(v any) any {
        |\tswitch x := v.(type) {
        |\tcase []byte:
        |\t\treturn string(x)
        |\tcase time.Time:
        |\t\treturn x.UTC().Format(time.RFC3339)
        |\tdefault:
        |\t\treturn v
        |\t}
        |}
        |
        |func queryAll(ctx context.Context, db *bun.DB, q string) ([]map[string]any, error) {
        |\trows, err := db.QueryContext(ctx, q)
        |\tif err != nil {
        |\t\treturn nil, err
        |\t}
        |\tdefer func() { _ = rows.Close() }()
        |\tcols, err := rows.Columns()
        |\tif err != nil {
        |\t\treturn nil, err
        |\t}
        |\tout := make([]map[string]any, 0)
        |\tfor rows.Next() {
        |\t\tvals := make([]any, len(cols))
        |\t\tptrs := make([]any, len(cols))
        |\t\tfor i := range vals {
        |\t\t\tptrs[i] = &vals[i]
        |\t\t}
        |\t\tif err := rows.Scan(ptrs...); err != nil {
        |\t\t\treturn nil, err
        |\t\t}
        |\t\tm := map[string]any{}
        |\t\tfor i, c := range cols {
        |\t\t\tm[c] = normalize(vals[i])
        |\t\t}
        |\t\tout = append(out, m)
        |\t}
        |\treturn out, rows.Err()
        |}
        |
        |func Register(r chi.Router, db *bun.DB, guard func(http.Handler) http.Handler) {
        |\tr.With(guard).Post("/admin/reset", func(w http.ResponseWriter, req *http.Request) {
        |\t\tctx := req.Context()
        |\t\t_ = ctx
        |$resetStmts
        |\t\tw.WriteHeader(204)
        |\t})
        |
        |\tr.With(guard).Get("/admin/state", func(w http.ResponseWriter, req *http.Request) {
        |\t\tctx := req.Context()
        |\t\t_ = ctx
        |\t\tstate := map[string]any{}
        |$stateAssigns
        |\t\tw.Header().Set("Content-Type", "application/json")
        |\t\tw.WriteHeader(200)
        |\t\t_ = json.NewEncoder(w).Encode(state)
        |\t})$seedSection
        |}
        |""".stripMargin

  private def seedHandler(e: entity_decl, isPg: Boolean): String =
    val snake  = Naming.toSnakeCase(entName(e))
    val pk     = AdminModel.primaryKeyField(e).getOrElse("id")
    val pkCol  = Naming.toColumnName(pk)
    val fields = entFields(e)
    val cols   = fields.map(f => (fldName(f), Naming.toColumnName(fldName(f))))
    val colChecks = cols
      .map: (_, col) =>
        s"""\t\tif v, ok := body[${GoLit.str(col)}]; ok {
           |\t\t\tcolNames = append(colNames, ${GoLit.str(col)})
           |\t\t\targs = append(args, v)
           |\t\t}""".stripMargin
      .mkString("\n")
    val insertExec =
      if isPg then
        s"""\t\tq := fmt.Sprintf("INSERT INTO ${Naming.toTableName(
            entName(e)
          )} (%s) VALUES (%s) RETURNING ${pkCol}", join(colNames), placeholders(len(args), true))
           |\t\tvar id any
           |\t\tif err := db.QueryRowContext(ctx, q, args...).Scan(&id); err != nil {
           |\t\t\thttp.Error(w, err.Error(), 500)
           |\t\t\treturn
           |\t\t}""".stripMargin
      else
        s"""\t\tq := fmt.Sprintf("INSERT INTO ${Naming.toTableName(
            entName(e)
          )} (%s) VALUES (%s)", join(colNames), placeholders(len(args), false))
           |\t\tres, err := db.ExecContext(ctx, q, args...)
           |\t\tif err != nil {
           |\t\t\thttp.Error(w, err.Error(), 500)
           |\t\t\treturn
           |\t\t}
           |\t\tid, _ := res.LastInsertId()""".stripMargin
    s"""|\tr.With(guard).Post(${GoLit.str(
         s"/admin/seed/$snake"
       )}, func(w http.ResponseWriter, req *http.Request) {
        |\t\tctx := req.Context()
        |\t\tvar body map[string]any
        |\t\tif err := json.NewDecoder(req.Body).Decode(&body); err != nil {
        |\t\t\thttp.Error(w, err.Error(), 400)
        |\t\t\treturn
        |\t\t}
        |\t\t_ = body
        |\t\tcolNames := make([]string, 0)
        |\t\targs := make([]any, 0)
        |$colChecks
        |$insertExec
        |\t\tw.Header().Set("Content-Type", "application/json")
        |\t\tw.WriteHeader(201)
        |\t\t_ = json.NewEncoder(w).Encode(map[string]any{${GoLit.str(pkCol)}: id})
        |\t})""".stripMargin